feat: backend V3

+++ backend/projeto-rancho/src/main/java/sptech/school/projeto_rancho/dto/auth/UserDTO.java
@@ -1,29 +1,23 @@
 package sptech.school.projeto_rancho.dto.auth;
 
-/**
- * Caminho: src/main/java/com/rancho/api/dto/auth/UserDTO.java
- *
- * ATENÇÃO: o campo é "name" (não "nome") pois o frontend
- * usa user.name para exibir o nome do utilizador logado.
- */
 public class UserDTO {
 
-    private Long id;
-    private String name;   // ← frontend lê user.name
+    private Integer id;
+    private String name;
     private String email;
     private String role;
 
     public UserDTO() {}
 
-    public UserDTO(Long id, String name, String email, String role) {
+    public UserDTO(Integer id, String name, String email, String role) {
         this.id    = id;
         this.name  = name;
         this.email = email;
         this.role  = role;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
